:recycle: refactor : make enums package and move
