--- src/main/java/com/run_us/server/domains/crew/domain/CrewJoinRequestStatusEnum.java
@@ -1,7 +1,0 @@
-package com.run_us.server.domains.crew.domain;
-
-public enum CrewJoinRequestStatusEnum {
-    WAIT,
-    APPROVE,
-    REJECT;
-}
