--- src/main/java/com/run_us/server/domains/crew/domain/CrewMembershipRoleEnum.java
@@ -1,5 +1,0 @@
-package com.run_us.server.domains.crew.domain;
-
-public enum CrewMembershipRoleEnum {
-    MEMBER;
-}
