+++ src/main/java/com/run_us/server/domains/crew/domain/enums/CrewJoinRequestStatusEnum.java
@@ -1,0 +1,7 @@
+package com.run_us.server.domains.crew.domain.enums;
+
+public enum CrewJoinRequestStatusEnum {
+    WAITING,
+    APPROVED,
+    REJECTED;
+}
