+++ src/main/java/com/run_us/server/domains/crew/domain/enums/CrewMembershipRoleEnum.java
@@ -1,0 +1,5 @@
+package com.run_us.server.domains.crew.domain.enums;
+
+public enum CrewMembershipRoleEnum {
+    MEMBER;
+}
