+++ src/main/java/com/run_us/server/domains/crew/domain/enums/CrewStatusEnum.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domains.crew.domain;
+package com.run_us.server.domains.crew.domain.enums;
 
 public enum CrewStatusEnum {
     ACTIVE,
